All Elevator Blocks, Scenes 1 and 2
Elevator blocks are what I call blocks that fall down, move up, left, or right when you stand on them.  They are all implemented and can vary the speed and delay time, and how long they last

Scenes 1 and 2 are done.  At the end of scene 1 there is a portal that takes you to scene 2 and changes your position.